Organize data into content-based graphs
Previously GOM used a graph per account and per resource but this
cannot be done in Tracker 3.

In Tracker 3, each graph is stored as a separate SQLite database
and currently a maximum of ten graphs can be included in a single
query due to SQLite's limitations.

The new layout is modelled after how tracker-miner-fs uses graphs:
we use a graph per content type (Contacts, Documents, Pictures),
plus an extra graph to store accounts and folders/collections.